requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
discord.py==2.3.2
distlib==0.3.6
et-xmlfile==1.1.0
filelock==3.10.0
frozenlist==1.4.1
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
multidict==6.0.4
numpy==1.25.0
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.0.1
platformdirs==3.1.1
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.21.0
yarl==1.9.4